mesh: Moved expr impl out of public headers.
